logisticsapp/logapp/requirements.txt
amqp==2.3.2
anyjson==0.3.3
argon2-cffi==16.3.0
arrow==0.10.0
asn1crypto==0.22.0
billiard==3.5.0.3
blessed==1.14.2
celery==4.1.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cheroot==5.7.0
cryptography==1.9
dj-database-url==0.5.0
Django==1.11
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.1.0
django-filter==1.0.4
django-oauth-toolkit==1.0.0
django-oauth2-provider==0.2.6.1
django-picklefield==1.0.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
et-xmlfile==1.0.1
eventlet==0.23.0
future==0.16.0
greenlet==0.4.13
gunicorn==19.7.1
idna==2.5
jdcal==1.3
kombu==4.2.1
Markdown==2.6.8
mysqlclient==1.3.10
oauthlib==2.0.2
olefile==0.44
openpyxl==2.4.9
Pillow==4.1.1
portend==2.1.2
pycparser==2.17
PyJWT==1.5.0
pyOpenSSL==17.0.0
PySocks==1.6.7
python-dateutil==2.6.0
python-decouple==3.1
python-mimeparse==1.6.0
pytz==2017.2
redis==2.10.5
requests==2.17.3
shortuuid==0.5.0
simplejson==3.13.2
six==1.10.0
sqlparse==0.2.3
tempora==1.8
twilio==6.4.1
urllib3==1.21.1
vine==1.1.4
waitress==1.1.0
wcwidth==0.1.7
whitenoise==3.3.1
